requirements.txt
autodocstring==0.1.2
autopep8==2.3.2
classopt==0.2.1
datasets==2.18.0
flake8==7.1.1
ipdb
ipykernel==6.29.5
jupyter==1.0.0
loralib==0.1.2
numpy==1.23.3
ninja==1.11.1.1
open_clip_torch==2.0.2
POT==0.8.2
pickle5==0.0.11
pillow==9.2.0
scikit-learn==1.4.1.post1
timm<=1.0.10
tqdm==4.64.1
transformers==4.46.3
wandb==0.17.4